requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0 